Added L3 ZTP module and updated playbooks
