+++ ansible/library/pn_l2_third_party.py
@@ -23,22 +23,13 @@ import shlex
 
 DOCUMENTATION = """
 ---
-module: pn_ztp
+module: pn_l2_third_party
 author: 'Pluribus Networks (@saptaktakalkar)'
-modified by: 'Pluribus Networks (@gauravbajaj)'
-version: 1
-short_description: CLI command to do zero touch provisioning.
+short_description: CLI command to configure layer2 zero touch provisioning.
 description:
     Zero Touch Provisioning (ZTP) allows you to provision new switches in your
-    network automatically, without manual intervention.
-    It performs following steps:
-        - Accept EULA
-        - Disable STP
-        - Enable all ports
-        - Create/Join fabric
-        - Auto configure vlags
-        - Change fabric type to in-band
-        - Enable STP
+    network automatically, without manual intervention. It will configure layer2
+    on all leaf switches, with spines being third party (non PN) switches.
 options:
     pn_cliusername:
         description:
@@ -50,24 +41,6 @@ options:
         - Provide login password if user is not root.
       required: False
       type: str
-    pn_fabric_name:
-      description:
-        - Specify name of the fabric.
-      required: False
-      type: str
-    pn_fabric_network:
-      description:
-        - Specify fabric network type as either mgmt or in-band.
-      required: False
-      type: str
-      choices: ['mgmt', 'in-band']
-      default: 'mgmt'
-    pn_run_l2:
-      description:
-        - Boolean flag to decide whether to configure vlag and link IPs or not.
-      required: False
-      type: bool
-      default: False
     pn_spine_list:
       description:
         - Specify list of Spine hosts.
@@ -84,78 +57,6 @@ options:
       required: False
       default: False
       type: bool
-    pn_inband_ip:
-      description:
-        - Inband ips to be assigned to switches starting with this value.
-      required: False
-      default: 172.16.0.0/24.
-      type: str
-    pn_fabric_control_network:
-      description:
-        - Specify fabric control network as either mgmt or in-band.
-      required: False
-      type: str
-      choices: ['mgmt', 'in-band']
-      default: 'mgmt'
-    pn_toggle_40g:
-      description:
-        - Flag to indicate if 40g ports should be converted to 10g ports or not.
-      required: False
-      default: True
-      type: bool
-    pn_current_switch:
-      description:
-        - Name of the switch on which this task is currently getting executed.
-      required: False
-      type: str
-    pn_static_setup:
-      description:
-        - Flag to indicate if static values should be assign to
-        following switch setup params.
-      required: False
-      default: False
-      type: bool
-    pn_mgmt_ip:
-      description:
-        - Specify MGMT-IP value to be assign if pn_static_setup is True.
-      required: False
-      type: str
-    pn_mgmt_ip_subnet:
-      description:
-        - Specify subnet mask for MGMT-IP value to be assign if
-        pn_static_setup is True.
-      required: False
-      type: str
-    pn_gateway_ip:
-      description:
-        - Specify GATEWAY-IP value to be assign if pn_static_setup is True.
-      required: False
-      type: str
-    pn_dns_ip:
-      description:
-        - Specify DNS-IP value to be assign if pn_static_setup is True.
-      required: False
-      type: str
-    pn_dns_secondary_ip:
-      description:
-        - Specify DNS-SECONDARY-IP value to be assign if pn_static_setup is True
-      required: False
-      type: str
-    pn_domain_name:
-      description:
-        - Specify DOMAIN-NAME value to be assign if pn_static_setup is True.
-      required: False
-      type: str
-    pn_ntp_server:
-      description:
-        - Specify NTP-SERVER value to be assign if pn_static_setup is True.
-      required: False
-      type: str
-    pn_web_api:
-      description:
-        - Flag to enable web api.
-      default: True
-      type: bool
     pn_stp:
       description:
         - Flag to enable STP at the end.
@@ -165,29 +66,25 @@ options:
 """
 
 EXAMPLES = """
-- name: Zero Touch Provisioning - Initial setup
-  pn_ztp:
+- name: Zero Touch Provisioning - Layer2 setup with 3rd party switches
+  pn_l2_third_party:
     pn_cliusername: "{{ USERNAME }}"
     pn_clipassword: "{{ PASSWORD }}"
-    pn_fabric_name: ztp-fabric
-    pn_run_l2: False
-
-- name: Zero Touch Provisioning - Layer2 setup
-  pn_ztp:
-    pn_cliusername: "{{ USERNAME }}"
-    pn_clipassword: "{{ PASSWORD }}"
-    pn_fabric_name: ztp-fabric
-    pn_fabric_type: layer3
-    pn_run_l2: True
+    pn_spine_list: "{{ groups['spine'] }}"
+    pn_leaf_list: "{{ groups['leaf'] }}"
 """
 
 RETURN = """
-msg:
+stdout:
   description: The set of responses for each command.
   returned: always
   type: str
 changed:
   description: Indicates whether the CLI caused changes on the target.
+  returned: always
+  type: bool
+failed:
+  description: Indicates whether or not the execution failed on the target.
   returned: always
   type: bool
 """
@@ -241,109 +138,6 @@ def run_cli(module, cli):
         return 'Success'
 
 
-def auto_accept_eula(module):
-    """
-    Method to accept the EULA when we first login to a new switch.
-    :param module: The Ansible module to fetch input parameters.
-    :return: The output of run_cli() method.
-    """
-    password = module.params['pn_clipassword']
-    cli = ' /usr/bin/cli --quiet --skip-setup eula-show '
-    cli = shlex.split(cli)
-    rc, out, err = module.run_command(cli)
-
-    if err:
-        cli = '/usr/bin/cli --quiet'
-        cli += ' --skip-setup --script-password '
-        cli += ' switch-setup-modify password ' + password
-        cli += ' eula-accepted true '
-        return run_cli(module, cli)
-    elif out:
-        return ' EULA has been accepted already! '
-
-
-def update_switch_names(module, switch_name):
-    """
-    Method to update switch names.
-    :param module: The Ansible module to fetch input parameters.
-    :param switch_name: Name to assign to the switch.
-    :return: String describing whether switch name got modified or not.
-    """
-    cli = pn_cli(module)
-    cli += ' switch-setup-show format switch-name '
-    if switch_name in run_cli(module, cli).split()[1]:
-        return ' Switch name is same as hostname! '
-    else:
-        cli = pn_cli(module)
-        cli += ' switch-setup-modify switch-name ' + switch_name
-        run_cli(module, cli)
-        return ' Updated switch name to match hostname! '
-
-
-def make_switch_setup_static(module):
-    """
-    Method to assign static values to different switch setup parameters.
-    :param module: The Ansible module to fetch input parameters.
-    """
-    mgmt_ip = module.params['pn_mgmt_ip']
-    mgmt_ip_subnet = module.params['pn_mgmt_ip_subnet']
-    gateway_ip = module.params['pn_gateway_ip']
-    dns_ip = module.params['pn_dns_ip']
-    dns_secondary_ip = module.params['pn_dns_secondary_ip']
-    domain_name = module.params['pn_domain_name']
-    ntp_server = module.params['pn_ntp_server']
-
-    if mgmt_ip:
-        cli = pn_cli(module)
-        ip = mgmt_ip + '/' + mgmt_ip_subnet
-        cli += ' switch-setup-modify mgmt-ip ' + ip
-        run_cli(module, cli)
-
-    if gateway_ip:
-        cli = pn_cli(module)
-        cli += ' switch-setup-modify gateway-ip ' + gateway_ip
-        run_cli(module, cli)
-
-    if dns_ip:
-        cli = pn_cli(module)
-        cli += ' switch-setup-modify dns-ip ' + dns_ip
-        run_cli(module, cli)
-
-    if dns_secondary_ip:
-        cli = pn_cli(module)
-        cli += ' switch-setup-modify dns-secondary-ip ' + dns_secondary_ip
-        run_cli(module, cli)
-
-    if domain_name:
-        cli = pn_cli(module)
-        cli += ' switch-setup-modify domain-name ' + domain_name
-        run_cli(module, cli)
-
-    if ntp_server:
-        cli = pn_cli(module)
-        cli += ' switch-setup-modify ntp-server ' + ntp_server
-        run_cli(module, cli)
-
-
-def modify_stp_local(module, modify_flag):
-    """
-    Method to enable/disable STP (Spanning Tree Protocol) on a switch.
-    :param module: The Ansible module to fetch input parameters.
-    :param modify_flag: Enable/disable flag to set.
-    :return: The output of run_cli() method.
-    """
-    cli = pn_cli(module)
-    cli += ' switch-local stp-show format enable '
-    current_state = run_cli(module, cli).split()[1]
-
-    if current_state == 'yes':
-        cli = pn_cli(module)
-        cli += ' stp-modify ' + modify_flag
-        return run_cli(module, cli)
-    else:
-        return ' STP is already disabled! '
-
-
 def modify_stp(module, modify_flag):
     """
     Method to enable/disable STP (Spanning Tree Protocol) on all switches.
@@ -365,116 +159,10 @@ def modify_stp(module, modify_flag):
             cli += ' switch ' + switch
             cli += ' stp-modify ' + modify_flag
             output += run_cli(module, cli)
-            output += ' '
         else:
             output += ' STP is already enabled! '
 
     return output
-
-
-def configure_control_network(module, network):
-    """
-    Method to configure the fabric control network.
-    :param module: The Ansible module to fetch input parameters.
-    :param network: It can be in-band or management.
-    :return: The output of run_cli() method.
-    """
-    cli = pn_cli(module)
-    cli += ' fabric-info format control-network '
-    current_control_network = run_cli(module, cli).split()[1]
-
-    if current_control_network == network:
-        return ' Fabric is already in %s control network ' % network
-    else:
-        cli = pn_cli(module)
-        cli += ' fabric-local-modify control-network ' + network
-        return run_cli(module, cli)
-
-
-def enable_ports(module):
-    """
-    Method to enable all ports of a switch.
-    :param module: The Ansible module to fetch input parameters.
-    :return: The output of run_cli() method.
-    """
-    cli = pn_cli(module)
-    cli += ' port-config-show format port no-show-headers '
-    out = run_cli(module, cli)
-
-    cli = pn_cli(module)
-    cli += ' port-config-show format port speed 40g no-show-headers '
-    out_40g = run_cli(module, cli)
-    out_remove10g = []
-
-    if len(out_40g) > 0 and out_40g != 'Success':
-        out_40g = out_40g.split()
-        out_40g = list(set(out_40g))
-        if len(out_40g) > 0:
-            for port_number in out_40g:
-                out_remove10g.append(str(int(port_number) + int(1)))
-                out_remove10g.append(str(int(port_number) + int(2)))
-                out_remove10g.append(str(int(port_number) + int(3)))
-
-    if out:
-        out = out.split()
-        out = set(out) - set(out_remove10g)
-        out = list(out)
-        if out:
-            ports = ','.join(out)
-            cli = pn_cli(module)
-            cli += ' port-config-modify port %s enable ' % ports
-            return run_cli(module, cli)
-    else:
-        return out
-
-
-def create_or_join_fabric(module, fabric_name, fabric_network):
-    """
-    Method to create/join a fabric with default fabric type as mgmt.
-    :param module: The Ansible module to fetch input parameters.
-    :param fabric_name: Name of the fabric to create/join.
-    :param fabric_network: Type of the fabric to create (mgmt/in-band).
-    Default value: mgmt
-    :return: The output of run_cli() method.
-    """
-    cli = pn_cli(module)
-    clicopy = cli
-
-    cli += ' fabric-show format name no-show-headers '
-    fabrics_names = run_cli(module, cli).split()
-
-    if fabric_name not in fabrics_names:
-        cli = clicopy
-        cli += ' fabric-create name ' + fabric_name
-        cli += ' fabric-network ' + fabric_network
-    else:
-        cli = clicopy
-        cli += ' fabric-info format name no-show-headers'
-        cli = shlex.split(cli)
-        rc, out, err = module.run_command(cli)
-
-        if err:
-            cli = clicopy
-            cli += ' fabric-join name ' + fabric_name
-        elif out:
-            present_fabric_name = out.split()
-            if present_fabric_name[1] not in fabrics_names:
-                cli = clicopy
-                cli += ' fabric-join name ' + fabric_name
-            else:
-                return 'Switch already in the fabric'
-
-    return run_cli(module, cli)
-
-
-def enable_web_api(module):
-    """
-    Method to enable web api on switches.
-    :param module: The Ansible module to fetch input parameters.
-    """
-    cli = pn_cli(module)
-    cli += ' admin-service-modify web if mgmt '
-    run_cli(module, cli)
 
 
 def update_fabric_network_to_inband(module):
@@ -734,99 +422,17 @@ def configure_auto_vlag(module):
     """
     spine_list = module.params['pn_spine_list']
     leaf_list = module.params['pn_leaf_list']
+
+    # Configure trunk, vlag for clustered leaf switches.
     non_clustered_leafs = find_non_clustered_leafs(module, leaf_list)
     output = configure_trunk_vlag_for_clustered_leafs(module,
                                                       non_clustered_leafs,
                                                       spine_list)
 
+    # Configure trunk, vlag for non clustered leaf switches.
     non_clustered_leafs = find_non_clustered_leafs(module, leaf_list)
     output += configure_trunk_non_clustered_leafs(module, non_clustered_leafs,
                                                   spine_list)
-    return output
-
-
-def toggle_40g_local(module):
-    """
-    Method to toggle 40g ports to 10g ports.
-    :param module: The Ansible module to fetch input parameters.
-    :return: The output of run_cli() method.
-    """
-    output = ''
-    cli = pn_cli(module)
-    clicopy = cli
-    cli += ' switch-local lldp-show format local-port no-show-headers '
-    local_ports = run_cli(module, cli).split()
-
-    cli = clicopy
-    cli += ' switch-local port-config-show speed 40g '
-    cli += ' format port no-show-headers '
-    ports_40g = run_cli(module, cli)
-    if len(ports_40g) > 0 and ports_40g != 'Success':
-        ports_40g = ports_40g.split()
-        ports_to_modify = list(set(ports_40g) - set(local_ports))
-
-        for port in ports_to_modify:
-            end_port = int(port) + 4
-            range_port = port + '-' + str(end_port)
-
-            cli = clicopy
-            cli += ' switch-local port-config-modify port %s ' % range_port
-            cli += ' speed disable '
-            output += 'port range_port ' + range_port + ' disabled'
-            output += run_cli(module, cli)
-
-            cli = clicopy
-            cli += ' switch-local port-config-modify port %s ' % range_port
-            cli += ' speed 10g '
-            output += 'port range_port ' + range_port + ' 10g converted'
-            output += run_cli(module, cli)
-
-            cli = clicopy
-            cli += ' switch-local port-config-modify port %s ' % range_port
-            cli += ' enable '
-            output += 'port range_port ' + range_port + '  enabled'
-            output += run_cli(module, cli)
-
-    return output
-
-
-def assign_inband_ip(module, inband_address):
-    """
-    Method to assign in-band ips to switches.
-    :param module: The Ansible module to fetch input parameters.
-    :param inband_address: The network ip for the in-band ips.
-    :return: String describing if in-band ips got assigned to switches or not.
-    """
-    output = ''
-    address = inband_address.split('.')
-    static_part = str(address[0]) + '.' + str(address[1]) + '.'
-    static_part += str(address[2]) + '.'
-    last_octet = str(address[3]).split('/')
-    subnet = last_octet[1]
-
-    cli = pn_cli(module)
-    clicopy = cli
-    cli += ' fabric-node-show format name no-show-headers '
-    switch_names = run_cli(module, cli).split()
-    switch_names.sort()
-
-    if len(switch_names) > 0:
-        ip_count = 1
-        if len(switch_names) + ip_count - 1 <= 255:
-            for switch in switch_names:
-                ip = static_part + str(ip_count) + '/' + subnet
-                cli = clicopy
-                cli += ' switch %s switch-setup-modify ' % switch
-                cli += ' in-band-ip ' + ip
-                if 'Success' in run_cli(module, cli):
-                    output += ' Assigned in-band ips to ' + switch
-
-                ip_count += 1
-        else:
-            output += ' Not enough in-band ips available for all the switches '
-    else:
-        output += ' No switches present '
-
     return output
 
 
@@ -836,142 +442,38 @@ def main():
         argument_spec=dict(
             pn_cliusername=dict(required=False, type='str'),
             pn_clipassword=dict(required=False, type='str', no_log=True),
-            pn_fabric_name=dict(required=True, type='str'),
-            pn_fabric_network=dict(required=False, type='str',
-                                   choices=['mgmt', 'in-band'],
-                                   default='mgmt'),
             pn_spine_list=dict(required=False, type='list'),
             pn_leaf_list=dict(required=False, type='list'),
-            pn_run_l2=dict(required=False, type='bool', default=False),
             pn_update_fabric_to_inband=dict(required=False, type='bool',
                                             default=False),
-            pn_inband_ip=dict(required=False, type='str',
-                              default='172.16.0.0/24'),
-            pn_fabric_control_network=dict(required=False, type='str',
-                                           choices=['mgmt', 'in-band'],
-                                           default='mgmt'),
-            pn_toggle_40g=dict(required=False, type='bool', default=True),
-            pn_current_switch=dict(required=False, type='str'),
-            pn_static_setup=dict(required=False, type='bool', default=False),
-            pn_mgmt_ip=dict(required=False, type='str'),
-            pn_mgmt_ip_subnet=dict(required=False, type='str'),
-            pn_gateway_ip=dict(required=False, type='str'),
-            pn_dns_ip=dict(required=False, type='str'),
-            pn_dns_secondary_ip=dict(required=False, type='str'),
-            pn_domain_name=dict(required=False, type='str'),
-            pn_ntp_server=dict(required=False, type='str'),
-            pn_web_api=dict(type='bool', default=True),
             pn_stp=dict(required=False, type='bool', default=False),
         )
     )
 
-    fabric_name = module.params['pn_fabric_name']
-    fabric_network = module.params['pn_fabric_network']
-    control_network = module.params['pn_fabric_control_network']
-    run_l2 = module.params['pn_run_l2']
-    update_fabric_to_inband = module.params['pn_update_fabric_to_inband']
-    inband_address = module.params['pn_inband_ip']
-    toggle_40g_flag = module.params['pn_toggle_40g']
-    current_switch = module.params['pn_current_switch']
-    message = ' '
     global CHANGED_FLAG
     CHANGED_FLAG = []
 
-    if not run_l2:
-        # Auto accept EULA
-        eula_out_msg = auto_accept_eula(module)
-        if 'Setup completed successfully' in eula_out_msg:
-            message += ' EULA accepted on %s! ' % current_switch
+    # L2 setup (auto-vlag)
+    message = configure_auto_vlag(module)
+
+    # Update fabric network to in-band if flag is True
+    if module.params['pn_update_fabric_to_inband']:
+        if 'Success' in update_fabric_network_to_inband(module):
+            message += ' Updated fabric network to in-band! '
             CHANGED_FLAG.append(True)
         else:
-            message += eula_out_msg
+            message += ' Fabric network is already in-band! '
             CHANGED_FLAG.append(False)
 
-        # Update switch names to match host names from hosts file
-        if 'Updated' in update_switch_names(module, current_switch):
-            CHANGED_FLAG.append(True)
-        else:
-            CHANGED_FLAG.append(False)
-
-        # Make switch setup static
-        if module.params['pn_static_setup']:
-            make_switch_setup_static(module)
-
-        # Create/join fabric
-        if 'already in the fabric' in create_or_join_fabric(module, fabric_name,
-                                                            fabric_network):
-            message += ' %s is already in fabric %s! ' % (current_switch,
-                                                          fabric_name)
-            CHANGED_FLAG.append(False)
-        else:
-            message += ' %s has joined fabric %s! ' % (current_switch,
-                                                       fabric_name)
-            CHANGED_FLAG.append(True)
-
-        # Configure fabric control network to either mgmt or in-band
-        net_out_msg = configure_control_network(module, control_network)
-        if 'Success' in net_out_msg:
-            message += ' Configured fabric control network to %s on %s! ' % (
-                control_network, current_switch)
-            CHANGED_FLAG.append(True)
-        else:
-            message += net_out_msg
-            CHANGED_FLAG.append(False)
-
-        # Enable web api if flag is True
-        if module.params['pn_web_api']:
-            enable_web_api(module)
-
-        # Disable STP
-        stp_out_msg = modify_stp_local(module, 'disable')
+    # Enable STP if flag is True
+    if module.params['pn_stp']:
+        stp_out_msg = modify_stp(module, 'enable')
         if 'Success' in stp_out_msg:
-            message += ' STP disabled on %s! ' % current_switch
+            message += ' STP enabled! '
             CHANGED_FLAG.append(True)
         else:
-            message += stp_out_msg
+            message += ' STP is already enabled! '
             CHANGED_FLAG.append(False)
-
-        # Enable ports
-        ports_out_msg = enable_ports(module)
-        if 'Success' in ports_out_msg:
-            message += ' Ports enabled on %s! ' % current_switch
-            CHANGED_FLAG.append(True)
-        else:
-            message += ports_out_msg
-            CHANGED_FLAG.append(False)
-
-        # Toggle 40g ports to 10g
-        if toggle_40g_flag:
-            if toggle_40g_local(module):
-                message += ' Toggled 40G ports to 10G on %s! ' % current_switch
-                CHANGED_FLAG.append(True)
-            else:
-                CHANGED_FLAG.append(False)
-    else:
-        # Assign in-band ips
-        message += assign_inband_ip(module, inband_address)
-
-        # L2 setup (auto-vlag)
-        message += configure_auto_vlag(module)
-
-        # Update fabric network to in-band if flag is True
-        if update_fabric_to_inband:
-            if 'Success' in update_fabric_network_to_inband(module):
-                message += ' Updated fabric network to in-band! '
-                CHANGED_FLAG.append(True)
-            else:
-                message += ' Fabric network is already in-band! '
-                CHANGED_FLAG.append(False)
-
-        # Enable STP if flag is True
-        if module.params['pn_stp']:
-            stp_out_msg = modify_stp(module, 'enable')
-            if 'Success' in stp_out_msg:
-                message += ' STP enabled! '
-                CHANGED_FLAG.append(True)
-            else:
-                message += ' STP is already enabled! '
-                CHANGED_FLAG.append(False)
 
     # Exit the module and return the required JSON
     module.exit_json(
