+++ ansible/playbooks/pn_l3_ztp.yml
@@ -1,0 +1,45 @@
+---
+
+
+# This task is to configure ZTP layer 3 setup.
+# It uses pn_l3_ztp.py module from library/ directory.
+# pn_cliusername and pn_clipassword comes from vars file - cli_vault.yml
+# If the tasks fails then it will retry as specified by retries count.
+- name: Zero Touch Provisioning - Layer3 setup
+  hosts: spine[0]
+  become: true
+  become_method: su
+  become_user: root
+
+  vars_files:
+  - cli_vault.yml
+
+  tasks:
+    - name: Auto configure link IPs
+      pn_l3_ztp:
+        pn_cliusername: "{{ USERNAME }}"        # Cli username (value comes from cli_vault.yml).
+        pn_clipassword: "{{ PASSWORD }}"        # Cli password (value comes from cli_vault.yml).
+        pn_spine_list: "{{ groups['spine'] }}"  # List of all spine switches mentioned under [spine] grp in hosts file.
+        pn_leaf_list: "{{ groups['leaf'] }}"    # List of all leaf switches mentioned under [leaf] grp in hosts file.
+        pn_net_address: '172.168.1.0'           # Network address required to calculate link IPs for layer3 fabric.
+        pn_cidr: '24'                           # Subnet mask required to calculate link IPs for layer3 fabric.
+        pn_supernet: '30'                       # Supernet mask required to calculate link IPs for layer3 fabric.
+        pn_assign_loopback: True                # Flag to indicate if loopback ips should be assigned to vrouters in layer3 fabric. Default: False.
+        # pn_loopback_ip: '109.109.109.0/24'    # Loopback ip value for vrouters in layer3 fabric. Default: 109.109.109.0/24.
+        # pn_bfd: True                          # Flag to indicate if BFD config should be added to vrouter interfaces in case of layer3 fabric. Default: False.
+        # pn_bfd_min_rx: 200                    # BFD-MIN-RX value required for adding BFD configuration to vrouter interfaces.
+        # pn_bfd_multiplier: 3                  # BFD_MULTIPLIER value required for adding BFD configuration to vrouter interfaces.
+        # pn_update_fabric_to_inband: False     # Flag to indicate if fabric network should be updated to in-band. Default: False.
+        # pn_stp: True                          # Flag to enable STP. Default: True.
+
+      register: ztp_l3_out                      # Variable to hold/register output of the above tasks.
+      until:  ztp_l3_out.failed != true         # If error pops up it will retry the code
+      retries: 3                                # This is the retries count
+      delay: 1
+      ignore_errors: yes                        # Flag to indicate if we should ignore errors if any.
+
+    - debug:
+        var: ztp_l3_out.stdout_lines            # Print stdout_lines of register variable.
+
+    - pause:
+        seconds: 2                              # Pause playbook execution for specified amount of time.
